Non-working commit
This is a NON-working commit. The next step from here on out is to
replace menu.dm, either with a new menu system or a type of form that
allows you to go back/forth somehow. Preferably a new menu system, since
really menu's should lead to forms, and be fairly 'static'.